Record all API requests in history and use double-precision timing
Move request history recording from individual handlers into
PreRouting/PostRouting so every API request is tracked with body,
headers, and status. Process/explorer handlers still add detailed
upstream call data. Change ResponseTimeMs from long to double across
all models, DB schemas, and frontend for two-decimal-place precision.
Fix History Entry line-wrap in Endpoint Explorer.

+++ src/Partio.Core/Database/Mysql/Queries/SetupQueries.cs
@@ -133,7 +133,7 @@ namespace Partio.Core.Database.Mysql.Queries
                 request_body_length BIGINT NULL,
                 response_body_length BIGINT NULL,
                 http_status INTEGER NULL,
-                response_time_ms BIGINT NULL,
+                response_time_ms DOUBLE NULL,
                 object_key VARCHAR(256) NULL,
                 created_utc TEXT NOT NULL,
                 completed_utc TEXT NULL
